+++ checks/evaluation/webhooks_test.go
@@ -34,7 +34,7 @@ func TestWebhooks(t *testing.T) {
 			name: "no webhooks",
 			findings: []finding.Finding{
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNotApplicable,
 				},
 			},
@@ -46,7 +46,7 @@ func TestWebhooks(t *testing.T) {
 			name: "1 webhook with no secret",
 			findings: []finding.Finding{
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 				},
 			},
@@ -58,7 +58,7 @@ func TestWebhooks(t *testing.T) {
 			name: "1 webhook with secret",
 			findings: []finding.Finding{
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomePositive,
 					Values: map[string]int{
 						"totalWebhooks": 1,
@@ -73,7 +73,7 @@ func TestWebhooks(t *testing.T) {
 			name: "2 webhooks one of which has secret",
 			findings: []finding.Finding{
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 2,
@@ -88,14 +88,14 @@ func TestWebhooks(t *testing.T) {
 			name: "Five webhooks three of which have secrets",
 			findings: []finding.Finding{
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 5,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 5,
@@ -110,7 +110,7 @@ func TestWebhooks(t *testing.T) {
 			name: "One of 12 webhooks does not have secrets",
 			findings: []finding.Finding{
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
@@ -125,84 +125,84 @@ func TestWebhooks(t *testing.T) {
 			name: "Score should not drop below min score",
 			findings: []finding.Finding{
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
 					},
 				},
 				{
-					Probe:   "webhooksWithoutSecret",
+					Probe:   "webhooksUseSecrets",
 					Outcome: finding.OutcomeNegative,
 					Values: map[string]int{
 						"totalWebhooks": 12,
